Adding the user menu code
A "user menu code" typically refers to a piece of computer programming code that is designed to create a menu or user interface for a software application. This code is responsible for displaying a list of options or actions that a user can choose from, and it often includes logic to respond to the user's selections.